Don't call Popen.wait() in LocalRunner
We already know the subprocess has exited because of poll()ing, so
calling wait() seems redundant.

This also removes the timeout arg from LocalRunner because it was only
being passed to wait(). I'm not sure if it would be useful to add this
back and then implement the timeout functionality in LocalRunner.