Rename package from taskrunner to runcommands
The latter name is available on PyPI. In addition, I think "command" may
be somewhat less ambiguous than "task"--i.e., "task" is often used to
refer to tasks/jobs that are being run via cron or the like, whereas
this package is about running commands from the command line.

This commit renames all instances of the word or substring "task" with
"command". In particular, the `task` decorator was renamed to `command`.
The default module where commands are found is now `commands.py` and the
default config file is `commands.cfg`.