[completion] Start from current word when locating command
Previously, we looked for the command to show completions for starting
from the last word (and going backwards), but that doesn't work when the
cursor is moved before a command. E.g., it didn't work for this:

    run --[CURSOR] local ls

because we were looking for completions for the local command instead of
the run script.

Now we ignore the current word and anything after it.