In get_default_prepend_path(), ignore bad asset paths
Not sure this is the *best* idea, but it follows what we were already
doing there with discarding paths that don't correspond to an existing
directory.

The basic idea is that every time the `local` command is called, it
recomputes the path and paths that didn't exist earlier may exist later.