Convert hide options to bool-or-string
This allows for `--hide` instead of `--hide all` and `--no-hide` instead
of `--hide none` while still supporting `--hide stdout`, etc.

The default value for all `hide` options was changed from `None` to
`False`.

This applies to the main script and the `local` and `remote` commands.
In the main script, this requires some special-casing to handle `--hide`
vs `--hide <value>`, which isn't so great.