Revamp/standardize injection of global defaults into command args
This applies to `echo` and `hide`. The global defaults for these (passed
as options to the main script) will be used when these options aren't
passed for individual commands.

If a global default isn't set, it won't be used now. This can happen if
a project's config sets `run.echo` or `run.hide` to `None`.