Pass arg type from command definition to Parameter
I.e., when a command is defined like so:

    @command(type={'bananas': list})
    def consume_bananas(config, bananas=None):
        # Consume bananas
        # If no bananas provided, cry instead
        if not bananas:
            print('Waaaaaah....')
        while bananas:
            bananas.pop()

the `Parameter` object created for the `bananas` arg will have its
`type` attribute set to `list` instead of deriving `type` from the
default value (`None` in this case).

This is especially useful for dict options where it's typical to see
a default specified as `None` instead of `{}`. This is less of an issue
with list-type options because `()` can be used as the default.

This change allows some logic to be simplified and made more consistent.