Fix command decorator's first arg
Rename it from `name_or_wrapped` to `name` so the name of a command can
be specified as `@command(name='xyz')` instead of having to specify the
name as a positional arg or using `name_or_wrapped='xyz'` (blech).

`name_or_wrapped` was intended as internal documentation of what the
first arg can be, but exposing a nice API here is more important.