Revamp handling of Ctrl-C in LocalRunner
Instead of immediately aborting on Ctrl-C, send a SIGINT to the
subprocess and let it decide what to do with it. After the subprocess
exits normally, then abort.

For interactive programs like less, this will require the user to press
q before the Ctrl-C will take effect.

This hopefully obviates the need for the `stty sane` kludge added in
39a85ef.