Automatically call replacement commands
That is, if a command is defined with the name of an existing, it will
replace the existing command and the replacement command will
automatically be called in place of the original command when the
original command is called directly.

This is intended to match the behavior of running commands from the
command line: only replacement commands are available via the CLI.

It's still possible to call an original command's implementation via its
`implementation()` method.

XXX: This behavior might be too magical.