Improve stdin handling in RemoteRunner
Ensure stdin is echoed back to the terminal immediately. When using a
PTY, this just means flushing stdout. When not using a PTY, this means
manually echoing stdin to stdout and then flushing.

Don't read stdin after remote command has exited. This seems pointless
and could possibly cause errors if data is sent to a process that has
already exited. This differs from reading stdout and stderr, because
their buffers might still contain data that needs to be read.