Revamp configuration of command args
Instead of passing the keyword arguments `help`, `type`, and `name`
to `Command` via the `@command` decorator to configure args, annotations
are now used.

The annotation for an arg can be either a type, a help string, or a
dict specifying any of `short_option`, `type`, `choices`, or `help`.

This reduces duplication and potential mistakes in matching arg names
when specifying config for an arg, and I think it's also a little less
unwieldy.

It also supports overriding short option names. (This commit only stores
overridden short options but doesn't do anything with them.)