Revamp completion
The `complete` command now takes the following args: the entire command
line, the current token, the cursor position, and a shell name.

We now handle tokenizing the command line ourselves instead of having
the shell pass in the tokens. The idea is that we should handle all
completion logic internally both for consistency and simplicity (this
makes the bash completion script simpler, for example).