Move local & remote commands to their respective runner modules
I.e., move the definition of the `local` command from `runners.commands`
to `runners.local` and the `remote` command to `runners.remote`.

This is so these commands can be imported separately. In particular, we
want to avoid importing the `runners.remote` module when Paramiko isn't
yet installed.

This is mainly intended for internal use; these commands are still
exported via `runners.commands` as well as `commands`.