Rewrite
YAML is now used for config instead of custom a INI/JSON format. This
reduces the amount of custom code we need to maintain. Interpoloation is
done via Jinja2, somewhat in the spirit of Ansible.

The config hierarchy now looks like this (lowest to highest):

- Defaults defined in config file.
- Keyword args.
- Globals defined in config file. Commands that have corresponding args
  will be passed the corresponding globals.
- Default args defined in config file. Passed to commands.
- Command line args and direct call args.

A global `config` object is no longer passed to all commands. This was
an idea copied from Fabric/Invoke, but I don't like it for a few
reasons:

- It's not explicit and can encourage dumping all kinds of stuff into
  the global `config` making it confusing sometimes to reason about
  where values come from.
- Not all commands need to access the global config, but they had to
  have the `config` arg regardless.
- It doesn't fit with the standalone console script use case.
- We could have used a "dependency injection" strategy to determine if
  a command should be passed the global config, but that seemed too
  magical.

Instead, commands should be explicit about which globals they're
interested by adding the corresponding args.

Note that it's possible to continue passing a global config object to
all commands by simply defining a global named `config` with the
relevant items. Alternatively, if a command has `**kwargs`, the global
config will be passed that way.

The `local` and `remote` runners have been simplified. `local` simply
wraps `subprocess.run()` and `remote` simply runs `ssh` via the `local`
command. All the stuff relating to streams was removed and Paramiko is
no longer a dependency.

Following from that, RunCommands is no longer trying to be
a full-fledged Fabric-esque deployment tool. There are other tools, such
as Ansible, Salt, etc, that do all that and more. RunCommands is now
focused on just being an easy way to create console scripts and
collections of commands that you'd use to drive something like Ansible
(e.g., instead of Bash scripts).

Python 3.4 and below are no longer supported. 3.3 is EOL and 3.4 doesn't
have `subprocess.run()`. TODO: Maybe add back support for 3.4.

In addition, numerous other improvements were made.