Add keep_slash flag to path utils
`util.path.abs_path` and `util.path.asset_path` have a new `keep_slash`
flag that will cause trailing slashes to be kept if present.

This is for use with the `sync` command, which uses `rsync` under the
hood, which treats directory paths with trailing slashes differently
than those without.