Convert yes values to tuple before inserting into prompt
In the `confirm()` utility function, when `yes_values` is a string, it
needs to be converted to a tuple before inserting the first item into
the prompt string.

Before, for example, `prompt_values='yes'` was displayed as "y" instead
of "yes" in the prompt, which was misleading since in this case "yes"
must be typed out and "y" won't work.