Collect values into tuples for args that are tuples
Previously, when an option was specified as a tuple (either explicitly
or because its default value was a tuple), corresponding values read
from the config file and/or the command line would be collected into
a list, which could be confusing or possibly cause errors inside
a command (e.g., when trying to concatenate a list to a tuple).

Now, corresponding command line args are directly collected into tuples
and default args read from the config file are converted to tuples when
appropriate.