Handle basic types for positional args
For example:

@command
def do_stuff(items: arg(type=tuple)):
    """Do stuff with things."""

Previously, this configuration would only allow a single value for
`items` on the command line and would also convert that single value to
a tuple of characters.

Now, this configuration will collection positional items into a dict,
list, or tuple when the type is specified as one of those.