Allow multiple short options to be grouped
This enables multiple short options to be passed to commands using
`-abc`.

This is achieved by a bit of pre-parsing of each command's `argv` in
`Command.parse_args()`. This is necessary because `allow_abbrev=False`
is set on the `ArgumentParser` instance created in `Command.arg_parser`.
The argparse docs say that `allow_abbrev` applies to long options, but
it also affects whether short options grouped behind a single dash will
be parsed into multiple short options.