Revamp handling of args, especially wrt. containers
Previously, when an arg was a container, values read from the command
line would be parsed as JSON automatically since there wasn't a way to
specify the type of the arg's values separately from its container type.

Now, when an arg is a container, its `type` is applied to the collected
values. An arg will be considered a container if A) it's explicitly
specified as being one using the new `container=<container type>` option
to the `arg` annotator or B) its default value is a container type.

To get the old behavior of parsing a container arg's values as JSON,
there's a new `json_value` type converter that can be specified as the
arg's `type`.

- Add `container` to `args.Arg` and `args.ArgConfig`.
- Add `nargs` to `args.Arg` and `args.ArgConfig`.
- Add `args.json_value()`.
- Remove `args.DictAddAction`.
- Remove `args.ListAppendAction`.
- Remove `args.NestedDictAddAction`.
- Remove `args.TupleAppendAction`.
- Remove `util.string.load_json_item()`.
- Remove `util.string.load_json_value()`.
- Move most of the logic/responsibility for determining `ArgParser` args
  from `Command.arg_parser` into the `Arg` class.
- Make `bool_or` class cache subclass created via `bool_or.__new__()`.
- Fix handling of `--`. Previously, this was used to divide the `run`
  console script's command line args from the remaining args (i.e., the
  commands to run and their args). The way it did so was simplistic and
  also didn't allow `--` to be used with individual commands, which is
  needed by commands that can be passed positional args that look like
  options. This has been fixed so that `--` will still split `run`'s
  args from the remaining, but this is now done in a more sensible way.
  In addition, each command can now use `--` to indicate the end of its
  options.

Builds on 2915afd4d1f4687bf193943eb4d9e7615811bf2b