Ensure last multi short option value when partitioning sys.argv
This applies only to the args for `run` when partitioning `sys.argv`.

When a multi short option group is found *and* the last option in the
group takes a value *and* a value wasn't provided at the end of the
group like -abc<value>, consume the value from `argv` and add it to
`run`'s arg list.

Amends 5dc4014
Refs 0d332f9