Include keyword-only args that have a default value
When converting a command's parameters to args, exclude only those
keyword-only args that don't have a default value instead of excluding
*all* keyword-only args.

This allows for commands defined like this:

    @command
    def test(*tests, fail_fast=True):
        if tests:
            "Run specific tests"
        else:
            "Run all tests"

I'm not entirely sure if keyword-only args that don't have a default
value should be excluded either. Does it make sense for commands to have
required args that are specified via --option? Are there scripts in the
wild that do that?