Add option for positional args w/ defaults set in config
Previously, a positional arg with a default set in the config file would
have its `nargs` set to * or ? (for container vs non-container types),
but that strategy doesn't work when `nargs` is set to collect N items
(e.g., `nargs=4`).

Instead, add an option for the arg (`--{arg.name}`) so that argparse
will treat it as optional without messing with `nargs`.