Allow commands to specify multiple callbacks
One possibility is that each command could have just one callback that
composes all the necessary callback functionality, and that's what was
initially implemented in a8b07e33377c.

But whether or not some specific callback code needs to run may be
conditional. As such, it's easier allow multiple callbacks that can be
added as needed, depending on conditions.

Otherwise, in more complex cases, the one callback would have to
implement branching, which could be messy.

In simple cases with a single callback, the only difference is that you
have to write `callbacks=[callback]` instead of `callback=callback`.

Amends a8b07e33377c