In printer utility, flush by default console is a TTY
The built-in print function doesn't flush by default, but in the
scenarios where RunCommands is used interactively, I think it's usually
the desired behavior.

So this passes `flush=True` to the built-in `print` function by default
when the console is a TTY. When the console isn't a TTY, the default
behavior is used (i.e., `flush=False`).