Improve how args are passed down from base commands
Instead of passing an arg down when a subcommand has the corresponding
*arg*, it's now passed down when a subcommand has the corresponding
*parameter*.

In addition, when a subcommand has a required keyword-only parameter,
the base command's default value will be passed down if the option
wasn't passed to the base command.

The primary reason for thse changes is to allow commands to have
required keyword-only parameters that can act as receivers for base
command args while not allowing those args to be passed (from the
command line) to the subcommand directly.

Note that required keyword-only parameters are already ignored when
figuring out what a command's args are.

For example, imagine a top-level script that has a "dry run" arg and a
a couple subcommands that may need to access it:

    @command
    def base(dry_run=True):
        pass

    @base.subcommand
    def sub(*, dry_run):
        pass

    @base.subcommand
    def other_sub(*, dry_run):
        pass

The idea is that if there are a lot of commands with a "dry run" option,
this will help ensure it's handled consistently. For example, it
probably doesn't make sense to pass `dry_run=True` to one subcommand
while passing `dry_run=False` to another--forcing it to be passed only
to the base command ensure there's a single value for all commands.

At the same time, the subcommands might need access to the "dry run"
option, perhaps to decide whether to print messages or take some other
kind of action.