Pass more data to command callbacks
Instead of passing just the command instance, also pass the result and
an `aborted` flag. The `aborted` flag will be set whenever a command is
aborted via Ctrl-C, `abort()`, `confirm()`, or whenever a `RunAborted`
exception is caught.

Note that all callbacks will receive the same `abort` value. In
particular, if a subcommand aborts, any base command callbacks will
receive `abort=True`.

This intended for use when cleaning up in a base command callback.