In printer utility, set default stream in method
That is, rather than setting the default stream using the keyword arg
`file=sys.stdout` for `print` and `file=sys.stderr` for the error-
printing methods, set `file=None` for all methods and grab the
appropriate stream dynamically inside each method.

This is so `sys.sterr` and `sys.stdout` can be redirected consistently.
For example, in tests.