Convert None & int results to Result objects
When a command returns either `None` or an int, a corresponding `Result`
object will now be created.

This will cause a `Result` object to be passed to callbacks in all cases
except where a command returns an object that's not `None`, an int, or
a `Result` (or a `Result`-like object).

This also changes the default return code from 1 to 0, as returning 0 on
non-error return/exit makes more sense (and I can't remember why 1 was
the default in the first place).