Set type from default value for container args w/o type
For container args that don't specify a type for the objects in the
container, use the type of the first item in the default value.

Note that this requires the default value for such args to be either an
empty container, a container with one or more items, `None`, or empty.