Wrap optional arg types to convert "" to None
This supports overriding an optional arg's default with `None` from the
command linelike so:

    run some-command --some-option=

When using this syntax, the value that gets passed to the optional arg's
type constructor is an empty string, which would previously cause an
error in many cases (e.g., for an int type arg).

Since passing an empty string doesn't seem like a common use case and
overriding a default with `None` seems more useful--i.e., I have
specific use cases for it--that's what this allows.

Also, in cases where an empty string *is* the desired value, that's easy
to work around by checking for `None`.