Add initial, simplistic make-like functionality
Commands can now specify `creates` (output paths/targets) to be created
from optional`sources` (source paths/prerequisites).

When *both* are specified, the source paths will be checked to see
whether they're newer than the output paths. If *any* are, *all* of the
output paths will be created.

When *only* output paths are specified, they'll *all* be created when
*any* of them doesn't already exist.

When source paths are specified but no matching files exist, a warning
will be printed (maybe should be a hard error?).

When output paths should be created but aren't, a warning will be
printed for each missing path.

Outstanding questions/issues:

- Should callbacks be run when no output paths are created?
- Add a `--force-create` flag to `run`?
- Add a `--clean` flag to `run`?
- Should output paths be `touch`ed automatically? Currently they're not,
  which means they may have to be `touch`ed manually (depending on the
  command).