Add util.is_mapping() and is_sequence()
Checking whether an object is a sequence is a pain because you always
have to make sure it's not a string. That's the impetus for adding
`is_sequence()`.

Since it's pretty common to check whether an object is a mapping or
sequence at the same time, that's the impetus for adding `is_mapping()`.

This makes these checks consistent and tidy.