Discover commands module & config file in CWD or parents
Previously, only the current directory was checked to find the commands
module and config file (if not specified). Now, we look in the current
directory and, if the files aren't found there, continue upward until
a project root is reached or the files system root is reached.

Note that for now commands still have to explicitly change directory if
they need to be run from the project root (see the `test` command for
example). It might make sense to automatically change to the directory
containing the commands module.

In addition, the commands module and config file are discovered
independently, but it might be better to find the config file based on
the location of the commands module (e.g., require it to be in the same
directory).