In remote command, handle sudo + stdout capture
This is a bit of a hack to handle the case where the `remote` command is
passed either `sudo` or `run_as` AND stdout is captured.

In this case, we need to hide the remote sudo prompt and show a fake
prompt locally. Otherwise, the remote prompt will be included in the
captured stdout.

TODO: See if there's a better way to handle this.